pyproject.toml
[project]
name = "rerun-visualization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "imagecodecs==2024.9.22",
    "numcodecs==0.13.1",
    "numpy==1.26.4",
    "pydantic==2.10.0",
    "rerun-sdk==0.21.0",
    "zarr==2.18.2",
]

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "I", "UP"]

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
